migrated component-based EventSim to Palladio nightly. * ProbeSpec has been completely deactivated for the moment * completely untested so far
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Core/branches/SimulationComponents@29235 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ edu.kit.ipd.sdq.eventsim.core/src/edu/kit/ipd/sdq/eventsim/resources/listener/IStateListener.java
@@ -20,6 +20,6 @@ public interface IStateListener {
      *            an identifier for the resource instance, for instance a CPU core. If the resource
      *            has only one instance, the instanceId must be 0 for that instance.
      */
-    public void stateChanged(int state, int instanceId);
+    public void stateChanged(long state, int instanceId);
 
 }
